More work on the ICC front: - Less strict warnings in debug-mode (still unusable, because ICC warns about _a lot_) - Explicit conversions - Clarify parameter evaluation order
git-svn-id: svn+ssh://svn.gna.org/svn/warzone/trunk@2550 4a71c877-e1ca-e34f-864e-861f7616d084

+++ lib/framework/trig.c
@@ -53,7 +53,7 @@ static float aSqrt[SQRT_ACCURACY];
 /* Initialise the Trig tables */
 BOOL trigInitialise(void)
 {
-	float val = 0.0, inc = 2.0 * M_PI / TRIG_DEGREES;
+	float val = 0.0f, inc = 2.0f * (float)M_PI / TRIG_DEGREES;
 	int i;
 
 	// Initialise the tables
@@ -64,12 +64,12 @@ BOOL trigInitialise(void)
 		val += inc;
 	}
 
-	inc = 2.0 / (TRIG_ACCURACY-1);
+	inc = 2.0f / (TRIG_ACCURACY-1);
 	val = -1;
 	for (i = 0; i < TRIG_ACCURACY; i++)
 	{
-		aInvSin[i] = asinf(val) * (float)TRIG_DEGREES / (2.0 * M_PI);
-		aInvCos[i] = acosf(val) * (float)TRIG_DEGREES / (2.0 * M_PI);
+		aInvSin[i] = asinf(val) * (float)TRIG_DEGREES / (2.0f * (float)M_PI);
+		aInvCos[i] = acosf(val) * (float)TRIG_DEGREES / (2.0f * (float)M_PI);
 		val += inc;
 	}
